Stock Opname: unify stock calc, decimal money, atomic numbering
Audit remediation for the Stock Opname / Barang Rusak slice:

- Unify the running-stock calculation behind a single query
  (BuildStockInfoQuery) so GetStockData (single item) and
  GetProductStockInfo (bulk import / year-end closing) agree on Qty
  System and HPP; remove the un-aggregated opening subquery that could
  return an arbitrary row. Source manual-add HPP from the same weighted
  average, falling back to product-master HPP only when no movement.
- Remove dead/incorrect GetStockAkhir (omitted damaged goods, double-
  counted opening) plus other dead members (GetItemBarang, repo
  GetDaftarBarang, StokOpname GetDaftarPembelianBarang, Edit/Cetak stubs).
- Use decimal instead of double for money in DTOStoctOpnameMaster/Detail.
- Make Insert_BarangRusak transactional and validated, allocating its
  number atomically from the shared STOCK_OPNAME counter instead of self-
  incrementing the legacy nomor_transaksi table (which could collide with
  a Stock Opname save). Drop the separate UpdateTransactionNumber call.
- Enable the duplicate-opname guard on the Barang Rusak barcode-scan path
  (previously only the manual-pick path checked).
- Migration 0008: advance POS_TRANSACTION_COUNTER to the max SO number
  across both POS_STOCKOPNAME and POS_BARANGRUSAK (idempotent).

+++ BackOffice/Model/DTOStoctOpnameDetail.cs
@@ -12,10 +12,10 @@ namespace BackOffice.Model
             public string NOMOR_SO { get; set; }
             public string KODE_BARANG { get; set; }
             public string PRODUCTNAME { get; set; }
-            public double JUMLAHSISTEM { get; set; }
-            public double JUMLAHFISIK { get; set; }
-            public double SELISIH { get; set; }
-            public double HPP { get; set; }
-            public double TOTAL { get; set; }
+            public decimal JUMLAHSISTEM { get; set; }
+            public decimal JUMLAHFISIK { get; set; }
+            public decimal SELISIH { get; set; }
+            public decimal HPP { get; set; }
+            public decimal TOTAL { get; set; }
     }
 }
