+++ BackOffice/Model/DTOStoctOpnameMaster.cs
@@ -11,7 +11,7 @@ namespace BackOffice.Model
             public string BULAN { get; set; }
             public string NOMOR_SO { get; set; }
             public DateTime TANGGAL { get; set; }
-        public double TOTAL { get; set; }
+        public decimal TOTAL { get; set; }
         public List<DTOStoctOpnameDetail> Details { get; set; } 
         }
 }
